♻️ refactor: 폴더, 메소드, 변수 이름 수정
