+++ docs/README.md
@@ -4,8 +4,7 @@
 
 ```bash
 📦src
- ┣ 📂app
- ┃ ┣ 📜App.js
+ ┣ 📂class
  ┃ ┗ 📜Cars.js
  ┣ 📂utils
  ┃ ┣ 📜inputValidator.js
@@ -13,6 +12,7 @@
  ┣ 📂view
  ┃ ┣ 📜InputView.js
  ┃ ┗ 📜OutputView.js
+ ┣ 📜App.js
  ┗ 📜index.js
 ```
 
@@ -39,7 +39,7 @@
 
 ## ➌ 시도 횟수만큼 자동차를 전진시킨다.
 - [x] 0~9 사이의 랜덤값을 생성하여 4 이상이면 전진 여부는 `true` → `isMove()`
-- [x] `isMove()`의 결과에 따라 자동차 전진 OR 정지 → `moveCars()`
+- [x] `isMove()`의 결과에 따라 자동차 전진 OR 정지 → `moveOrStop()`
 - [x] 한 차수가 끝나면 모든 자동차의 이동 거리 출력 → `printCars()`
 
 ## ➍ 최종 우승자를 출력한다.
